create own files for polynomial and ComplexNumber class
