+++ NNPTPZ1/Program.cs
@@ -13,7 +13,7 @@ using System.Drawing.Text;
 using System.Drawing.Drawing2D;
 using System.Linq.Expressions;
 using System.Threading;
-using NNPTPZ1.Mathematics;
+using NNPTPZ1;
 
 namespace NNPTPZ1
 {
@@ -230,181 +230,4 @@ namespace NNPTPZ1
         }
     }
 
-    namespace Mathematics
-    {
-        public class Polynomial
-        {
-            /// <summary>
-            /// Coe
-            /// </summary>
-            public List<ComplexNumber> Coefficients { get; set; }
-
-            /// <summary>
-            /// Constructor
-            /// </summary>
-            public Polynomial() => Coefficients = new List<ComplexNumber>();
-
-            public void Add(ComplexNumber coefficient) =>
-                Coefficients.Add(coefficient);
-
-            /// <summary>
-            /// Derives this polynomial and creates new one
-            /// </summary>
-            /// <returns>Derivated polynomial</returns>
-            public Polynomial Derivative()
-            {
-                Polynomial derivatedPolynomial = new Polynomial();
-                for (int q = 1; q < Coefficients.Count; q++)
-                {
-                    derivatedPolynomial.Coefficients.Add(Coefficients[q].Multiply(new ComplexNumber() { Real = q }));
-                }
-
-                return derivatedPolynomial;
-            }
-
-            /// <summary>
-            /// Evaluates polynomial at given point
-            /// </summary>
-            /// <param name="x">point of evaluation</param>
-            /// <returns>y</returns>
-            public ComplexNumber Evaluate(double x)
-            {
-                var y = Evaluate(new ComplexNumber() { Real = x, Imaginary = 0 });
-                return y;
-            }
-
-            /// <summary>
-            /// Evaluates polynomial at given point
-            /// </summary>
-            /// <param name="x">point of evaluation</param>
-            /// <returns>y</returns>
-            public ComplexNumber Evaluate(ComplexNumber x)
-            {
-                ComplexNumber zero = ComplexNumber.Zero;
-                for (int i = 0; i < Coefficients.Count; i++)
-                {
-                    ComplexNumber coefficient = Coefficients[i];
-                    ComplexNumber currentTerm = x;
-                    int power = i;
-
-                    if (i > 0)
-                    {
-                        for (int j = 0; j < power - 1; j++)
-                            currentTerm = currentTerm.Multiply(x);
-
-                        coefficient = coefficient.Multiply(currentTerm);
-                    }
-
-                    zero = zero.Add(coefficient);
-                }
-
-                return zero;
-            }
-
-            /// <summary>
-            /// ToString
-            /// </summary>
-            /// <returns>String repr of polynomial</returns>
-            public override string ToString()
-            {
-                string termStrings = "";
-                int i = 0;
-                for (; i < Coefficients.Count; i++)
-                {
-                    termStrings += Coefficients[i];
-                    if (i > 0)
-                    {
-                        int j = 0;
-                        for (; j < i; j++)
-                        {
-                            termStrings += "x";
-                        }
-                    }
-                    if (i+1<Coefficients.Count)
-                    termStrings += " + ";
-                }
-                return termStrings;
-            }
-        }
-
-        public class ComplexNumber
-        {
-            public double Real { get; set; }
-            public float Imaginary { get; set; }
-
-            public override bool Equals(object obj)
-            {
-                if (obj is ComplexNumber)
-                {
-                    ComplexNumber x = obj as ComplexNumber;
-                    return x.Real == Real && x.Imaginary == Imaginary;
-                }
-                return base.Equals(obj);
-            }
-
-            public readonly static ComplexNumber Zero = new ComplexNumber()
-            {
-                Real = 0,
-                Imaginary = 0
-            };
-
-            public ComplexNumber Multiply(ComplexNumber complexNumber)
-            {
-                ComplexNumber dividedNumber = this;
-                // aRe*bRe + aRe*bIm*i + aIm*bRe*i + aIm*bIm*i*i
-                return new ComplexNumber()
-                {
-                    Real = dividedNumber.Real * complexNumber.Real - dividedNumber.Imaginary * complexNumber.Imaginary,
-                    Imaginary = (float)(dividedNumber.Real * complexNumber.Imaginary + dividedNumber.Imaginary * complexNumber.Real)
-                };
-            }
-            public double GetAbS()
-            {
-                return Math.Sqrt( Real * Real + Imaginary * Imaginary);
-            }
-
-            public ComplexNumber Add(ComplexNumber complexNumber)
-            {
-                ComplexNumber dividedNumber = this;
-                return new ComplexNumber()
-                {
-                    Real = dividedNumber.Real + complexNumber.Real,
-                    Imaginary = dividedNumber.Imaginary + complexNumber.Imaginary
-                };
-            }
-            public double GetAngleInDegrees()
-            {
-                return Math.Atan(Imaginary / Real);
-            }
-            public ComplexNumber Subtract(ComplexNumber complexNumber)
-            {
-                ComplexNumber dividedNumber = this;
-                return new ComplexNumber()
-                {
-                    Real = dividedNumber.Real - complexNumber.Real,
-                    Imaginary = dividedNumber.Imaginary - complexNumber.Imaginary
-                };
-            }
-
-            public override string ToString()
-            {
-                return $"({Real} + {Imaginary}i)";
-            }
-
-            internal ComplexNumber Divide(ComplexNumber complexNumber)
-            {
-                // (aRe + aIm*i) / (bRe + bIm*i)
-                // ((aRe + aIm*i) * (bRe - bIm*i)) / ((bRe + bIm*i) * (bRe - bIm*i))
-                //  bRe*bRe - bIm*bIm*i*i
-                var dividend = this.Multiply(new ComplexNumber() { Real = complexNumber.Real, Imaginary = -complexNumber.Imaginary });
-                var devider = complexNumber.Real * complexNumber.Real + complexNumber.Imaginary * complexNumber.Imaginary;
-
-                return new ComplexNumber()
-                {
-                    Real = dividend.Real / devider,
-                    Imaginary = (float)(dividend.Imaginary / devider)
-                };
-            }
-        }
-    }
 }
